testing

+++ lib/ParameterChecks.groovy
@@ -6,7 +6,7 @@ class ParameterChecks {
         assert !(params.GxE && !params.SNPs), "Please specify path to variants with --SNPs parameter when running --GxE"
         assert !(params.noCpG && params.noCHG && params.noCHH), "Please specify at least one methylation context for analysis!"
         assert params.coverage instanceof Integer && params.coverage >= 0, "Input coverage filter must be a non-negative integer!"
-        assert Float.valueOf(params.proportion) in 0..1, "Proportion of shared regions must be a decimal in the range of 0 and 1!"
+        assert Double.valueOf(params.proportion) in 0..1, "Proportion of shared regions must be a decimal in the range of 0 and 1!"
         assert params.input_FDR instanceof Float && params.input_FDR in 0..1, "Input FDR filter must be a decimal in the range of 0 and 1!"
         assert params.output_FDR instanceof Float && params.output_FDR in 0..1, "Output FDR filter must be a decimal in the range of 0 and 1!"
         assert params.Emodel_pv instanceof Float && params.Emodel_pv in 0..1, "Emodel p-value must be a decimal in the range of 0 and 1!"
